obsolete/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
requests==2.32.3
biopython==1.85
pydantic==2.8.2
